Add formatter_test.py and update formatter.MatchLine()
formatter.MatchLine() now checks if a regex found data and returns the data when true. With the pay rates regex at the front of the list, this will enhance performance in addition to guaranteeing results when a match is found. However, the USDA number and the DoD ID are sometimes on the same line and only one would be found through this implementation.